Random Chicken, CreateEntity() evolved & NPCs can show in different stages easily now
The chicken is generated randomly in the map now. NPCs are added to a list in Scene and Stages modules, with an ID and the stage where they should appear, which is used in the PostUpdate() of the Scene module. Their position can be set in the CreateEntity function. This function also allow to create diferent types of NPCs now. Finally, some unnecessary variables have been deleted and basic textures and animations of children entities have been moved to parent entities when it was possible.